:recycle: 클래스 네이밍 변경
