+++ YourName/YourName/Sources/Data Layer/Network/Protocols/Networking.swift
@@ -8,6 +8,6 @@
 import Foundation
 import RxSwift
 
-protocol Network {
+protocol Networking {
     func response<Api: API>(of api: Api) -> Observable<Api.Response>
 }
