+++ YourName/YourNameTests/NetworkTests/Test Doubles/MockAPI.swift
@@ -8,13 +8,6 @@
 @testable import YourName
 import Foundation
 
-struct TestResponse: Decodable {
-    var test: Int
-}
-extension TestResponse {
-    static var stub: TestResponse = TestResponse(test: 0)
-}
-
 final class MockAPI: API {
   typealias Response = TestResponse
